pydantic-ai-example/requirements.txt
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.2.2
google-auth==2.37.0
griffe==1.5.4
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.8.2
jsonpath-python==1.0.6
logfire-api==2.11.0
mistralai==1.2.5
mypy-extensions==1.0.0
openai==1.58.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic-ai==0.0.15
pydantic-ai-slim==0.0.15
pydantic_core==2.27.2
pydotenv==0.0.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
